refactor: 댓글 관련 DTO 네이밍 및 패키지 변경

+++ space-d/src/main/java/com/dnd/spaced/core/comment/domain/Comment.java
@@ -33,7 +33,7 @@ public class Comment extends BaseTimeEntity {
 
     private String content;
 
-    private int likeCount = 0;
+    private long likeCount = 0L;
 
     private boolean isDeleted = false;
